:recycle: Refactor useLikeUpdate Hook
#193 useLikeUpdate를 hook으로 분리
